pyproject.toml
[tool.poetry]
name = "FinTrack"
version = "0.1.0"
description = "Personal Finance Management App with Django, HTMX, Tailwind and Plaid"
authors = ["earthcomfy <hanagetbel93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
django-htmx = "^1.19.0"
django-tailwind = "^3.8.0"
django-allauth = { extras = ["mfa"], version = "^64.1.0" }
django-cotton = "^0.9.32"
plaid-python = "^24.0.0"
celery = {extras = ["redis"], version = "^5.4.0"}
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pre-commit = "^3.8.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.1"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.1"
pytest-django = "^4.8.0"
django-browser-reload = "^1.15.0"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
pytest-celery = "^1.1.1"

[build-system]
requires = ["poetry-config"]
build-backend = "poetry.config.masonry.api"

[tool.pytest.ini_options]
addopts = "--ds=django_finance.config.settings.test --nomigrations --cov=. --cov-report=html --reuse-db"
python_files = ["test_*.py"]
markers = [
    "unit: tests that are isolated from the db, external api calls and other mockable internal code.",
]

[tool.coverage.run]
include = ["apps/**"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    '*/.venv/*',
    '*/venv/*',
    '*/env/*',
    '*/staticfiles/*',
    '*/config/*',
    '*/__init__.py',
    '*apps.py',
    '*urls.py',
    '*wsgi.py',
    '*asgi.py',
    '*manage.py',
    '*requirements.txt',
]

[tool.ruff]
line-length = 119

[tool.djlint]
profile = "django"
ignore = "H031"
